refactor: #9 新規メモ作成処理をReducer内で完結できるように修正。
