+++ bear-clone/src/app/shared/actions/note.actions.ts
@@ -1,10 +1,9 @@
 import { createActionGroup, emptyProps, props } from "@ngrx/store";
-import { Note } from "../models/note.model";
 
 export const NoteActions = createActionGroup({
   source: "Note",
   events: {
-    "Add Notes": props<Note>(),
+    "Add Notes": emptyProps(),
     "Remove Notes": props<{ noteId: string }>(),
   },
 });
